OPENBUS-3076 Refatoração da classes abstratas OpenBusAction e InputDialog
- remoção das classes sem uso do framework do logistic
- simplificação da construção de ações e diálogos já com suporte embutido das strings de internacionalização
- correção das mensagens dos diálogos de remoção (contrato, provedor, consumidor e integração)

+++ BusAdminLib/src/main/java/tecgraf/openbus/admin/BusAdminImpl.java
@@ -74,7 +74,7 @@ public class BusAdminImpl implements BusAdmin {
    * @param port Porta do barramento
    * @param orb ORB do barramento
    */
-  public void getAdminFacets(String host, int port, ORB orb) {
+  public void getAdminFacets(String host, short port, ORB orb) {
     org.omg.CORBA.Object ref = buildCorbaLoc(host, port, orb);
     obtainRegistries(ref);
   }
@@ -414,7 +414,7 @@ public class BusAdminImpl implements BusAdmin {
   /**
    * {@inheritDoc}
    */
-  private static org.omg.CORBA.Object buildCorbaLoc(String host, int port,
+  private static org.omg.CORBA.Object buildCorbaLoc(String host, short port,
     ORB orb) {
     String str =
       String.format("corbaloc::1.0@%s:%d/%s", host, port, BusObjectKey.value);
